feat: Refactor commands for slash launch helper and token tracking
- Refactored slash launch helper and token tracking commands to use specific functions for command creation to improve code readability and maintainability.
