+++ src/track/track_handlers.go
@@ -3,11 +3,8 @@ package track
 import (
 	"log"
 	"slices"
-	"strings"
-	"time"
 
 	"github.com/bwmarrin/discordgo"
-	"github.com/xhit/go-str2duration/v2"
 )
 
 // getTokenValComponents returns the components for the token value
@@ -230,90 +227,6 @@ func HandleTokenDetails(s *discordgo.Session, i *discordgo.InteractionCreate) {
 	m.Components = &comp
 	m.SetContent(str)
 	s.ChannelMessageEditComplex(m)
-}
-
-// HandleTokenCommand will handle the /token command
-func HandleTokenCommand(s *discordgo.Session, i *discordgo.InteractionCreate) {
-	// User interacting with bot, is this first time ?
-	options := i.ApplicationCommandData().Options
-	optionMap := make(map[string]*discordgo.ApplicationCommandInteractionDataOption, len(options))
-	for _, opt := range options {
-		optionMap[opt.Name] = opt
-	}
-	channelID := i.ChannelID
-	linked := true
-	linkReceived := true
-	var duration time.Duration
-	if opt, ok := optionMap["duration"]; ok {
-		var err error
-		// Timespan of the contract duration
-		contractTimespan := strings.TrimSpace(opt.StringValue())
-		contractTimespan = strings.Replace(contractTimespan, "day", "d", -1)
-		contractTimespan = strings.Replace(contractTimespan, "hr", "h", -1)
-		contractTimespan = strings.Replace(contractTimespan, "min", "m", -1)
-		contractTimespan = strings.Replace(contractTimespan, "sec", "s", -1)
-		duration, err = str2duration.ParseDuration(contractTimespan)
-		if err != nil {
-			// Invalid duration, just assigning a 12h
-			duration = 12 * time.Hour
-		}
-	}
-	var trackingName = ""
-	if opt, ok := optionMap["name"]; ok {
-		trackingName = strings.TrimSpace(opt.StringValue())
-	}
-	if opt, ok := optionMap["linked"]; ok {
-		linked = opt.BoolValue()
-	}
-	/*
-		if opt, ok := optionMap["link-received"]; ok {
-			linkReceived = opt.BoolValue()
-		}
-	*/
-	if opt, ok := optionMap["contract-channel"]; ok {
-		input := strings.TrimSpace(opt.StringValue())
-		s := strings.Split(input, "/")
-		if len(s) > 0 {
-			// set channelID to last entry in the slice
-			channelID = s[len(s)-1]
-		}
-		linked = false
-		linkReceived = false
-	}
-	// Call into boost module to do that calculations
-	var userID string
-	if i.GuildID != "" {
-		userID = i.Member.User.ID
-	} else {
-		userID = i.User.ID
-	}
-
-	str, err := tokenTracking(s, channelID, userID, trackingName, duration, linked, linkReceived)
-
-	if err != nil {
-		str = err.Error()
-	} else {
-		var data discordgo.MessageSend
-		data.Content = str
-		data.Components = getTokenValComponents(false, trackingName) // Initial state
-
-		u, _ := s.UserChannelCreate(userID)
-		msg, _ := s.ChannelMessageSendComplex(u.ID, &data)
-		Tokens[userID].Coop[trackingName].TokenMessageID = msg.ID
-		Tokens[userID].Coop[trackingName].UserChannelID = u.ID
-
-		str += "Interact with the bot on " + u.Mention() + " to track your token values."
-	}
-
-	s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-		Type: discordgo.InteractionResponseChannelMessageWithSource,
-		Data: &discordgo.InteractionResponseData{
-			Content: str,
-			Flags:   discordgo.MessageFlagsEphemeral,
-		},
-	},
-	)
-	saveData(Tokens)
 }
 
 // HandleTokenComplete will close the token tracking
